Restructure the module
