--- compressed.hf
@@ -1,1 +1,0 @@
-YÏX
