+++ resources/input1.txt
