+++ resources/input2.txt
