+++ resources/input3.txt
