+++ resources/input4.txt
