+++ src/configs/Bit.java
@@ -1,4 +1,4 @@
-package com.company;
+package configs;
 
 public enum Bit {
     ZERO,
