+++ src/huffman/CompressionResult.java
@@ -1,4 +1,6 @@
-package com.company;
+package huffman;
+
+import utils.Utils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
