+++ src/utils/Utils.java
@@ -1,4 +1,4 @@
-package com.company;
+package utils;
 
 import java.util.ArrayList;
